requirements.txt
Flask==3.0.3
torch==2.2.2
torchvision==0.17.2
open-clip-torch==2.23.0
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.4
safetensors==0.4.3
